pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autoso"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "llama-index-core>=0.12.46",
    "llama-index-llms-anthropic>=0.7.6",
    "llama-index-llms-ollama>=0.4.0",
    "llama-index-embeddings-huggingface>=0.5.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "chromadb>=0.6.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "openai-whisper>=20231117",
    "yt-dlp>=2024.1.0",
    "python-docx>=1.1.0",
    "pydub>=0.25.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
include = ["autoso*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests that hit real external services (opt-in with --run-integration)",
]
